Tests the lobby activity without outside depedencies
